Open source non stock binary 120hz
Fixes 120hz mode being enabled by default and the setting doing the opposite of what it says (60hz mode toggle instead of 120hz mode toggle)

Also less stock binaries to worry about so leaner base. However this will require patches for the option to be exposed to the user.

These were added to 18.1+, however they are easily backported to 17.1 and seemingly work just fine here.
https://github.com/LineageOS/android_packages_apps_Settings/commit/28945c48f6142d4a7cdd812db041d94227e1f52a
https://github.com/LineageOS/android_packages_apps_Settings/commit/0a9477fa342901b98e07fb85f3df0a7717749971

I would like to also get the refresh rate overlay working here instead of relying on testufo and my eyes to tell the refresh rate, but that will be for later.